@@ -1,31 +1,35 @@
 {
   "name": "tourze/symfony-aop-pool-bundle",
-  "description": "AOP连接池",
+  "description": "Symfony AOP connection pool bundle for automatic resource pooling",
   "type": "symfony-bundle",
   "license": "MIT",
   "require": {
-    "php": "^8.1",
     "doctrine/dbal": "^4.0",
     "monolog/monolog": "^3.1",
     "psr/log": "^3|^2|^1",
-    "symfony/config": "^6.4",
-    "symfony/console": "^6.4",
-    "symfony/dependency-injection": "^6.4",
-    "symfony/event-dispatcher": "^6.4",
-    "symfony/framework-bundle": "^6.4",
-    "symfony/http-kernel": "^6.4",
-    "symfony/messenger": "^6.4",
-    "symfony/service-contracts": "^3.5",
-    "symfony/yaml": "^6.4 || ^7.1",
+    "symfony/config": "^7.3",
+    "symfony/console": "^7.3",
+    "symfony/dependency-injection": "^7.3",
+    "symfony/event-dispatcher": "^7.3",
+    "symfony/framework-bundle": "^7.3",
+    "symfony/http-kernel": "^7.3",
+    "symfony/messenger": "^7.3",
+    "symfony/service-contracts": "^3.6",
+    "symfony/yaml": "^7.3",
     "tourze/backtrace-helper": "0.1.*",
+    "tourze/bundle-dependency": "1.*",
     "tourze/php-pools": "0.0.*",
     "tourze/symfony-aop-bundle": "~0.0.4",
+    "tourze/symfony-dependency-service-loader": "0.0.*",
     "tourze/symfony-runtime-context-bundle": "0.0.*"
   },
   "require-dev": {
-    "maglnet/composer-require-checker": "^4",
     "phpstan/phpstan": "^2.1",
-    "phpunit/phpunit": "^10.0"
+    "phpunit/phpunit": "^11.5",
+    "tourze/phpunit-base": "1.*",
+    "tourze/phpunit-symfony-kernel-test": "0.0.*",
+    "tourze/phpunit-symfony-unit-test": "0.0.*",
+    "tourze/symfony-testing-framework": "0.0.*"
   },
   "autoload": {
     "psr-4": {
