@@ -2,34 +2,40 @@
 
 namespace Tourze\Symfony\AopPoolBundle\Tests\Service;
 
-use PHPUnit\Framework\TestCase;
+use Doctrine\DBAL\Driver\Statement;
+use Doctrine\DBAL\Result;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\RunTestsInSeparateProcesses;
+use Tourze\PHPUnitSymfonyKernelTest\AbstractIntegrationTestCase;
 use Tourze\Symfony\AopPoolBundle\Service\ConnectionLifecycleHandler;
 use Utopia\Pools\Connection;
 
-class ConnectionLifecycleHandlerTest extends TestCase
+/**
+ * @internal
+ */
+#[CoversClass(ConnectionLifecycleHandler::class)]
+#[RunTestsInSeparateProcesses]
+final class ConnectionLifecycleHandlerTest extends AbstractIntegrationTestCase
 {
     private ConnectionLifecycleHandler $handler;
 
-    protected function setUp(): void
+    protected function onSetUp(): void
     {
-        parent::setUp();
-
         // 设置环境变量
         $_ENV['SERVICE_POOL_CONNECTION_LIFETIME'] = '30';
 
         // 创建处理器实例
-        $this->handler = new ConnectionLifecycleHandler();
-    }
-
-    protected function tearDown(): void
-    {
-        // 清除环境变量
-        unset($_ENV['SERVICE_POOL_CONNECTION_LIFETIME']);
-        parent::tearDown();
+        $this->handler = self::getService(ConnectionLifecycleHandler::class);
     }
 
     public function testRegisterConnection(): void
     {
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们只需要一个占位符对象来测试 ConnectionLifecycleHandler 的行为
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
 
         // 调用注册方法
@@ -41,22 +47,34 @@ class ConnectionLifecycleHandlerTest extends TestCase
         $startTimes = $reflection->getValue($this->handler);
 
         $id = $this->handler->getConnectionId($connection);
-        $this->assertArrayHasKey($id, $startTimes);
+        self::assertArrayHasKey($id, $startTimes);
     }
 
     public function testGetConnectionId(): void
     {
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们只需要一个占位符对象来测试 ConnectionLifecycleHandler 的行为
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
 
         // 获取连接ID
         $id = $this->handler->getConnectionId($connection);
 
         // 验证ID与spl_object_hash一致
-        $this->assertEquals(spl_object_hash($connection), $id);
+        self::assertEquals(spl_object_hash($connection), $id);
     }
 
     public function testUnregisterConnection(): void
     {
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们只需要一个占位符对象来测试 ConnectionLifecycleHandler 的行为
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
 
         // 先注册连接
@@ -73,15 +91,21 @@ class ConnectionLifecycleHandlerTest extends TestCase
         $reflection->setAccessible(true);
         $startTimes = $reflection->getValue($this->handler);
 
-        $this->assertArrayNotHasKey($id, $startTimes);
+        self::assertArrayNotHasKey($id, $startTimes);
     }
 
     public function testGetConnectionAge(): void
     {
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们只需要一个占位符对象来测试 ConnectionLifecycleHandler 的行为
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
 
         // 获取未注册连接的年龄
-        $this->assertNull($this->handler->getConnectionAge($connection));
+        self::assertNull($this->handler->getConnectionAge($connection));
 
         // 注册连接
         $this->handler->registerConnection($connection);
@@ -90,11 +114,17 @@ class ConnectionLifecycleHandlerTest extends TestCase
         $age = $this->handler->getConnectionAge($connection);
 
         // 验证年龄为0或1（考虑到执行时间）
-        $this->assertLessThanOrEqual(1, $age);
+        self::assertLessThanOrEqual(1, $age);
     }
 
     public function testCheckConnectionExpired(): void
     {
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们只需要一个占位符对象来测试 ConnectionLifecycleHandler 的行为
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
 
         // 获取连接ID
@@ -109,19 +139,28 @@ class ConnectionLifecycleHandlerTest extends TestCase
 
         // 期望抛出异常
         $this->expectException(\Exception::class);
-        $this->expectExceptionMessage("连接已过期");
+        $this->expectExceptionMessage('连接已过期');
 
         $this->handler->checkConnection($connection);
     }
 
     public function testCheckRedisConnection(): void
     {
-        // 跳过测试如果 Redis 类不存在
-        if (!class_exists(\Redis::class)) {
-            $this->markTestSkipped('Redis class not found');
-        }
+        // Redis 测试使用 Mock 对象，不依赖真实的 Redis 扩展
 
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们需要模拟 getResource() 方法来返回 Redis 实例
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
+        /**
+         * 必须使用具体类 \Redis 而不是接口，因为：
+         * 1. PHP Redis 扩展没有提供接口，只提供了具体类
+         * 2. 我们需要模拟 ping() 方法来测试连接健康检查
+         * 3. 这是合理的，因为 Redis 扩展是一个底层的 PHP 扩展，而不是一个遵循接口设计原则的库
+         */
         $redis = $this->createMock(\Redis::class);
 
         // 设置返回Redis资源
@@ -133,7 +172,8 @@ class ConnectionLifecycleHandlerTest extends TestCase
         // 模拟ping成功
         $redis->expects($this->once())
             ->method('ping')
-            ->willReturn('+PONG');
+            ->willReturn('+PONG')
+        ;
 
         // 调用检查方法
         $this->handler->registerConnection($connection);
@@ -142,12 +182,21 @@ class ConnectionLifecycleHandlerTest extends TestCase
 
     public function testCheckRedisConnectionFailure(): void
     {
-        // 跳过测试如果 Redis 类不存在
-        if (!class_exists(\Redis::class)) {
-            $this->markTestSkipped('Redis class not found');
-        }
+        // Redis 测试使用 Mock 对象，不依赖真实的 Redis 扩展
 
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们需要模拟 getResource() 方法来返回 Redis 实例
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
+        /**
+         * 必须使用具体类 \Redis 而不是接口，因为：
+         * 1. PHP Redis 扩展没有提供接口，只提供了具体类
+         * 2. 我们需要模拟 ping() 方法来测试连接健康检查
+         * 3. 这是合理的，因为 Redis 扩展是一个底层的 PHP 扩展，而不是一个遵循接口设计原则的库
+         */
         $redis = $this->createMock(\Redis::class);
 
         // 设置返回Redis资源
@@ -159,37 +208,60 @@ class ConnectionLifecycleHandlerTest extends TestCase
         // 模拟ping失败
         $redis->expects($this->once())
             ->method('ping')
-            ->willThrowException(new \Exception('Connection lost'));
+            ->willThrowException(new \Exception('Connection lost'))
+        ;
 
         // 调用注册方法
         $this->handler->registerConnection($connection);
 
         // 期望抛出异常
         $this->expectException(\Exception::class);
-        $this->expectExceptionMessage("Redis连接不健康");
+        $this->expectExceptionMessage('Redis连接不健康');
 
         $this->handler->checkConnection($connection);
     }
 
     public function testCheckDatabaseConnection(): void
     {
-        // 跳过测试如果 Doctrine\DBAL\Connection 类不存在
-        if (!class_exists(\Doctrine\DBAL\Connection::class)) {
-            $this->markTestSkipped('Doctrine\DBAL\Connection class not found');
-        }
+        // DBAL 测试使用 Mock 对象，不依赖真实的 Doctrine DBAL
 
         // 测试需要模拟 Doctrine\DBAL\Result
         $result = null;
-        if (class_exists(\Doctrine\DBAL\Result::class)) {
-            $result = $this->createMock(\Doctrine\DBAL\Result::class);
+        if (class_exists(Result::class)) {
+            /**
+             * 必须使用具体类 \Doctrine\DBAL\Result 而不是接口，因为：
+             * 1. Doctrine DBAL 在不同版本中提供了不同的结果类，未统一接口
+             * 2. 我们需要模拟 fetchOne() 方法来测试查询结果
+             * 3. 这是合理的，因为这是为了兼容不同版本的 Doctrine DBAL
+             */
+            $result = $this->createMock(Result::class);
             $result->method('fetchOne')->willReturn('1');
         } else {
             // 适配旧版API
-            $result = $this->createMock(\Doctrine\DBAL\Driver\Statement::class);
+            /**
+             * 必须使用具体类 \Doctrine\DBAL\Driver\Statement 而不是接口，因为：
+             * 1. 这是旧版 Doctrine DBAL 的 API，在新版中已被 Result 类替代
+             * 2. 我们需要模拟 fetchColumn() 方法来测试查询结果
+             * 3. 这是合理的，因为这是为了保持向后兼容性
+             */
+            $result = $this->createMock(Statement::class);
             $result->method('fetchColumn')->willReturn('1');
         }
 
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们需要模拟 getResource() 方法来返回 DBAL Connection 实例
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
+        /**
+         * 必须使用具体类 \Doctrine\DBAL\Connection 而不是接口，因为：
+         * 1. 虽然 Doctrine 提供了 ConnectionInterface，但在测试中使用具体类更常见
+         * 2. 我们需要模拟 executeQuery() 方法来测试数据库连接健康检查
+         * 3. 这是可以接受的，但更好的做法是使用 Doctrine\DBAL\Driver\Connection 接口
+         * 替代方案：$this->createMock(\Doctrine\DBAL\Driver\Connection::class)
+         */
         $dbal = $this->createMock(\Doctrine\DBAL\Connection::class);
 
         // 设置返回DBAL资源
@@ -202,7 +274,8 @@ class ConnectionLifecycleHandlerTest extends TestCase
         $dbal->expects($this->once())
             ->method('executeQuery')
             ->with('SELECT 1')
-            ->willReturn($result);
+            ->willReturn($result)
+        ;
 
         // 调用检查方法
         $this->handler->registerConnection($connection);
@@ -211,12 +284,22 @@ class ConnectionLifecycleHandlerTest extends TestCase
 
     public function testCheckDatabaseConnectionFailure(): void
     {
-        // 跳过测试如果 Doctrine\DBAL\Connection 类不存在
-        if (!class_exists(\Doctrine\DBAL\Connection::class)) {
-            $this->markTestSkipped('Doctrine\DBAL\Connection class not found');
-        }
+        // DBAL 测试使用 Mock 对象，不依赖真实的 Doctrine DBAL
 
+        /**
+         * 必须使用具体类 Connection 而不是接口，因为：
+         * 1. Connection 是 Utopia\Pools 库提供的具体类，该库未提供相应接口
+         * 2. 我们需要模拟 getResource() 方法来返回 DBAL Connection 实例
+         * 3. 这是合理的，因为第三方库的设计不在我们的控制范围内
+         */
         $connection = $this->createMock(Connection::class);
+        /**
+         * 必须使用具体类 \Doctrine\DBAL\Connection 而不是接口，因为：
+         * 1. 虽然 Doctrine 提供了 ConnectionInterface，但在测试中使用具体类更常见
+         * 2. 我们需要模拟 executeQuery() 方法来测试数据库连接健康检查
+         * 3. 这是可以接受的，但更好的做法是使用 Doctrine\DBAL\Driver\Connection 接口
+         * 替代方案：$this->createMock(\Doctrine\DBAL\Driver\Connection::class)
+         */
         $dbal = $this->createMock(\Doctrine\DBAL\Connection::class);
 
         // 设置返回DBAL资源
@@ -229,14 +312,15 @@ class ConnectionLifecycleHandlerTest extends TestCase
         $dbal->expects($this->once())
             ->method('executeQuery')
             ->with('SELECT 1')
-            ->willThrowException(new \Exception('Connection lost'));
+            ->willThrowException(new \Exception('Connection lost'))
+        ;
 
         // 调用注册方法
         $this->handler->registerConnection($connection);
 
         // 期望抛出异常
         $this->expectException(\Exception::class);
-        $this->expectExceptionMessage("数据库连接不健康");
+        $this->expectExceptionMessage('数据库连接不健康');
 
         $this->handler->checkConnection($connection);
     }
